# Use prefix_same_as_start to avoid iteration in FindNextUserEntryInternal.