This avoids excessive iteration in tombstone fields.