@petermattis 2016-04-28
No new tests leaves me a little uncomfortable, but this is a performance change, not a correctness one. I'm not familiar with any of the RocksDB benchmarking infrastructure. I'd be happy to write a benchmark for this if someone provided guidance.