@igorcanadi 2016-04-28
Nice find!

Couple of comments:
1. It would be great to add a test. I just checked db_iter_test.cc and it has a nice mocking framework for testing db_iter. You could create an artificial test that will trigger the new code-path (where you stop the iteration early) and count the number of keys that were read from the underlying iterator. You can then assert that the number of keys read is as expected.
2. db_iter is not very well documented, especially given how important and complicated is the code. Can you please add one comment describing the prefix_check parameter in FindNextUserEntry* functions?
3. [optional] Would be interesting to quote any performance benefits you get in your benchmark in the commit message.

BTW, have you checked out SingleDelete() operation? It's a trick we use in MyRocks and MongoRocks to optimize thombstone removal for secondary keys. Unfortunately we don't have a good blog post on it (yet), but you can check out the commit message: https://reviews.facebook.net/D43179