@petermattis 2016-04-28
Ok, I'll take a look at `db_iter_test.cc` and add a bit of comment documentation to `db_iter.cc`. 

One of our benchmarks test doing `seek(key), put(key), delete(key)` with random keys. Previously, we were using `ReadOptions.iterate_upper_bound` to avoid pathological performance in `FindNextUserEntryInternal`, but we want to switch to `ReadOptions.prefix_same_as_start` so that we can reuse the same iterator across different keys. The "old" numbers below are when we switch to using `ReadOptions.prefix_same_as_start` without this change to RocksDB.

```
name                     old time/op  new time/op  delta
MVCCPutDelete_RocksDB-8   149µs ± 1%    10µs ± 2%  -93.42%  (p=0.000 n=10+9)
```

PS I don't see how the travis failure is related to this change and don't see what the appveyor failure is.