@petermattis 2016-04-28
And yes, I have looked at `SingleDelete`, though I'm not sure how we could use it safely yet.