@igorcanadi 2016-04-28
It doesn't look like travis failure is related. Appveyor is broken in the master as well (it's the windows build)