@petermattis 2016-04-28
@igorcanadi This is ready for another look. I made a small addition to `PrefixTest.PrefixValid` which fails before the `db_iter.cc` change and succeeds afterward. And I've added some documentation about the `prefix_check` parameter.