@igorcanadi 2016-04-28
Thanks, this looks great! I'll let @siying take another look and merge.